+++ src/components/Detail/menuDummy.json
@@ -1,0 +1,23 @@
+{
+  "id": 1,
+  "franchise": "스타벅스",
+  "imageUrl": "https://via.placeholder.com/300x150",
+  "title": "슈렉 프라프치노",
+  "originalTitle": "프라푸치노",
+  "author": {
+    "id": 111,
+    "name": "계란이 조아",
+    "profileImageUrl": "https://via.placeholder.com/300x150"
+  },
+  "content": "정말 맛이 있습니다",
+  "options": [
+    { "name": "에스프레소샷", "description": "1" },
+    { "name": "자바칩", "description": "약간" },
+    { "name": "아이스크림", "description": "한스쿱" }
+  ],
+  "expectedPrice": 5000,
+  "tastes": ["매콤함", "느끼함"],
+  "likes": 100,
+  "createdAt": "!12312123",
+  "updatedAt": "!@31231213"
+}
